+++ src/math/bcknd/device/ax_helm_svv_KS_full_device.F90
@@ -30,8 +30,8 @@
 ! ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 ! POSSIBILITY OF SUCH DAMAGE.
 !
-module ax_helm_svv_full_device
-  use ax_helm_svv_full, only : ax_helm_svv_full_t
+module ax_helm_svv_KS_full_device
+  use ax_helm_svv_KS_full, only : ax_helm_svv_KS_full_t
   use num_types, only : rp
   use coefs, only : coef_t
   use space, only : space_t
@@ -43,22 +43,22 @@ module ax_helm_svv_full_device
   implicit none
   private
 
-  !> Device asymmetric full-stress SVV Helmholtz operator.
-  type, public, extends(ax_helm_svv_full_t) :: ax_helm_svv_full_device_t
+  !> Device Kirby-Sherwin full-stress SVV Helmholtz operator.
+  type, public, extends(ax_helm_svv_KS_full_t) :: ax_helm_svv_KS_full_device_t
    contains
      !> Compute the coupled vector product.
      procedure, pass(this) :: compute_vector => &
-          ax_helm_svv_full_device_compute_vector
-  end type ax_helm_svv_full_device_t
+          ax_helm_svv_KS_full_device_compute_vector
+  end type ax_helm_svv_KS_full_device_t
 
 #ifdef HAVE_HIP
   interface
-     subroutine hip_ax_helm_svv_full(au_d, av_d, aw_d, u_d, v_d, w_d, &
+     subroutine hip_ax_helm_svv_KS_full(au_d, av_d, aw_d, u_d, v_d, w_d, &
           dx_d, dy_d, dz_d, h1_d, &
           drdx_d, drdy_d, drdz_d, dsdx_d, dsdy_d, dsdz_d, &
           dtdx_d, dtdy_d, dtdz_d, jacinv_d, w3_d, svv_h1_d, &
           filter_r_d, filter_s_d, filter_t_d, nelv, lx) &
-          bind(c, name='hip_ax_helm_svv_full')
+          bind(c, name='hip_ax_helm_svv_KS_full')
        use, intrinsic :: iso_c_binding, only : c_ptr, c_int
        type(c_ptr), value :: au_d, av_d, aw_d
        type(c_ptr), value :: u_d, v_d, w_d
@@ -69,16 +69,16 @@ module ax_helm_svv_full_device
        type(c_ptr), value :: jacinv_d, w3_d, svv_h1_d
        type(c_ptr), value :: filter_r_d, filter_s_d, filter_t_d
        integer(c_int) :: nelv, lx
-     end subroutine hip_ax_helm_svv_full
+     end subroutine hip_ax_helm_svv_KS_full
   end interface
 #elif HAVE_CUDA
   interface
-     subroutine cuda_ax_helm_svv_full(au_d, av_d, aw_d, u_d, v_d, w_d, &
+     subroutine cuda_ax_helm_svv_KS_full(au_d, av_d, aw_d, u_d, v_d, w_d, &
           dx_d, dy_d, dz_d, h1_d, &
           drdx_d, drdy_d, drdz_d, dsdx_d, dsdy_d, dsdz_d, &
           dtdx_d, dtdy_d, dtdz_d, jacinv_d, w3_d, svv_h1_d, &
           filter_r_d, filter_s_d, filter_t_d, nelv, lx) &
-          bind(c, name='cuda_ax_helm_svv_full')
+          bind(c, name='cuda_ax_helm_svv_KS_full')
        use, intrinsic :: iso_c_binding, only : c_ptr, c_int
        type(c_ptr), value :: au_d, av_d, aw_d
        type(c_ptr), value :: u_d, v_d, w_d
@@ -89,14 +89,14 @@ module ax_helm_svv_full_device
        type(c_ptr), value :: jacinv_d, w3_d, svv_h1_d
        type(c_ptr), value :: filter_r_d, filter_s_d, filter_t_d
        integer(c_int) :: nelv, lx
-     end subroutine cuda_ax_helm_svv_full
+     end subroutine cuda_ax_helm_svv_KS_full
   end interface
 #endif
 
 contains
 
-  !> Compute the coupled asymmetric full-stress SVV Helmholtz product.
-  !! @param this Device asymmetric full-stress SVV operator.
+  !> Compute the coupled Kirby-Sherwin full-stress SVV Helmholtz product.
+  !! @param this Device Kirby-Sherwin full-stress SVV operator.
   !! @param au Result for the first component.
   !! @param av Result for the second component.
   !! @param aw Result for the third component.
@@ -106,9 +106,9 @@ contains
   !! @param coef Coefficients.
   !! @param msh Mesh.
   !! @param Xh Function space.
-  subroutine ax_helm_svv_full_device_compute_vector(this, au, av, aw, &
+  subroutine ax_helm_svv_KS_full_device_compute_vector(this, au, av, aw, &
        u, v, w, coef, msh, Xh)
-    class(ax_helm_svv_full_device_t), intent(in) :: this
+    class(ax_helm_svv_KS_full_device_t), intent(in) :: this
     type(mesh_t), intent(in) :: msh
     type(space_t), intent(in) :: Xh
     type(coef_t), intent(in) :: coef
@@ -147,7 +147,7 @@ contains
     end if
 
 #ifdef HAVE_HIP
-    call hip_ax_helm_svv_full(au_d, av_d, aw_d, u_d, v_d, w_d, &
+    call hip_ax_helm_svv_KS_full(au_d, av_d, aw_d, u_d, v_d, w_d, &
          Xh%dx_d, Xh%dy_d, Xh%dz_d, coef%h1_d, &
          coef%drdx_d, coef%drdy_d, coef%drdz_d, &
          coef%dsdx_d, coef%dsdy_d, coef%dsdz_d, &
@@ -155,7 +155,7 @@ contains
          Xh%w3_d, this%svv%h1_d, filter_r_d, filter_s_d, filter_t_d, &
          msh%nelv, Xh%lx)
 #elif HAVE_CUDA
-    call cuda_ax_helm_svv_full(au_d, av_d, aw_d, u_d, v_d, w_d, &
+    call cuda_ax_helm_svv_KS_full(au_d, av_d, aw_d, u_d, v_d, w_d, &
          Xh%dx_d, Xh%dy_d, Xh%dz_d, coef%h1_d, &
          coef%drdx_d, coef%drdy_d, coef%drdz_d, &
          coef%dsdx_d, coef%dsdy_d, coef%dsdz_d, &
@@ -172,6 +172,6 @@ contains
        call device_addcol4(aw_d, coef%h2_d, coef%B_d, w_d, coef%dof%size())
     end if
 
-  end subroutine ax_helm_svv_full_device_compute_vector
+  end subroutine ax_helm_svv_KS_full_device_compute_vector
 
-end module ax_helm_svv_full_device
+end module ax_helm_svv_KS_full_device